Split system and add friendly fire kills
