+++ Resources/Locale/en-US/_RMC14/game-presets/distress-signal-preset.ftl
@@ -57,7 +57,9 @@ rmc-distress-signal-round-stat-marine-projectiles-fired-top = Projectiles Fired:
 rmc-distress-signal-round-stat-marine-projectile-hits = Projectile Hits: {$count}
 rmc-distress-signal-round-stat-marine-projectile-hits-top = Projectile Hits: {$count} | {$name}: {$value}
 rmc-distress-signal-round-stat-friendly-fire-incidents = Friendly Fire Incidents: {$count}
-rmc-distress-signal-round-stat-friendly-fire-incidents-top = Friendly Fire Incidents: {$count} | {$name}: {$value}
+rmc-distress-signal-round-stat-friendly-fire-incidents-top = Friendly Fire Instances: {$count} | {$name}: {$value}
+rmc-distress-signal-round-stat-friendly-fire-kills = Friendly Fire Kills: {$count}
+rmc-distress-signal-round-stat-friendly-fire-kills-top = Friendly Fire Kills: {$count} | {$name}: {$value}
 rmc-distress-signal-round-stat-larva-extractions = Larva Extractions: {$count}
 rmc-distress-signal-round-stat-larva-extractions-top = Larva Extractions: {$count} | {$name}: {$value}
 rmc-distress-signal-round-stat-used-requisitions-budget = Requisitions Budget Used: {$count}
